frontend token auth

+++ app/frontend/utils/function.go
@@ -6,5 +6,5 @@ func GetUserIdFromCtx(ctx context.Context) uint32 {
 	if ctx.Value(UserIdKey) == nil {
 		return 0
 	}
-	return uint32(ctx.Value(UserIdKey).(float64))
+	return uint32(ctx.Value(UserIdKey).(int32))
 }
